move SipHash to CryptoContext
