+++ core-test/src/test/scala/org/bitcoins/core/gcs/GolombFilterTest.scala
@@ -2,7 +2,7 @@ package org.bitcoins.core.gcs
 
 import org.bitcoins.core.number.UInt64
 import org.bitcoins.core.protocol.CompactSizeUInt
-import org.bitcoins.crypto.DoubleSha256Digest
+import org.bitcoins.crypto.{DoubleSha256Digest, SipHashKey}
 import org.bitcoins.testkit.core.gen.CryptoGenerators._
 import org.bitcoins.testkit.core.gen.NumberGenerator
 import org.bitcoins.testkit.util.BitcoinSAsyncTest
